pyproject.toml
[tool.poetry]
name = "plantmon"
version = "0.1.0"
description = ""
authors = ["TranBinhMinh_Finn"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.4"
mysqlclient = "^2.1.1"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = {extras = ["crypto"], version = "^5.2.2"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

